Stage 3 core libraries. Thread/Event system complete.
Mutilthreaded locks still need to be implemented, but threads are complete and provide facilities for cancellation and abnormal termination.

Events have been completed, but WaitForEvent/Events still needs to be built.

Fixed errors in CastGuid.cpp and in NetRequest.cpp due to renamed types.

Added INetTcpSocket interface for the beginning of TCP communication.